+++ core/vm/arch/win32/memory.h
@@ -66,7 +66,7 @@ struct StackFrameMonitor {
 
 // holders
 struct CollectionInfo {
-  long* op_stack;
+  size_t* op_stack;
   long stack_pos;
 };
 
@@ -197,7 +197,7 @@ public:
   }
 
   // mark and sweep functions
-  static void CollectAllMemory(long* op_stack, long stack_pos);
+  static void CollectAllMemory(size_t* op_stack, long stack_pos);
   static uintptr_t WINAPI CollectMemory(LPVOID arg);
   static uintptr_t WINAPI CheckStatic(LPVOID arg);
   static uintptr_t WINAPI CheckStack(LPVOID arg);
@@ -221,7 +221,7 @@ public:
   static void CheckObject(long* mem, bool is_obj, const long depth);
 
   // memory allocation
-  static long* AllocateObject(const wchar_t* obj_name, long* op_stack, long stack_pos, bool collect = true) {
+  static long* AllocateObject(const wchar_t* obj_name, size_t* op_stack, long stack_pos, bool collect = true) {
     StackClass* cls = prgm->GetClass(obj_name);
     if(cls) {
       return AllocateObject(cls->GetId(), op_stack, stack_pos, collect);
@@ -233,8 +233,8 @@ public:
   //
   // object and array allocation
   //
-  static long* AllocateObject(const long obj_id, long* op_stack, long stack_pos, bool collect = true);
-  static long* AllocateArray(const long size, const MemoryType type, long* op_stack, long stack_pos, bool collect = true);
+  static long* AllocateObject(const long obj_id, size_t* op_stack, long stack_pos, bool collect = true);
+  static long* AllocateArray(const long size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
 
   //
   // object verification
